clean code (#288)
